Generalize EVMLA functions
