+++ src/context/function/mod.rs
@@ -2,5 +2,7 @@
 //! The common LLVM function entities.
 //!
 
+pub mod block;
 pub mod declaration;
+pub mod evmla_data;
 pub mod r#return;
