+++ src/context/traits/mod.rs
@@ -4,3 +4,4 @@
 
 pub mod address_space;
 pub mod evmla_data;
+pub mod evmla_function;
