+++ src/lib.rs
@@ -10,14 +10,18 @@ pub(crate) mod optimizer;
 pub(crate) mod target_machine;
 
 pub use self::context::attribute::Attribute;
-pub use self::context::block_key::BlockKey;
 pub use self::context::code_type::CodeType;
+pub use self::context::function::block::evmla_data::EVMLAData as FunctionBlockEVMLAData;
+pub use self::context::function::block::key::Key as BlockKey;
+pub use self::context::function::block::Block as FunctionBlock;
 pub use self::context::function::declaration::Declaration as FunctionDeclaration;
+pub use self::context::function::evmla_data::EVMLAData as FunctionEVMLAData;
 pub use self::context::function::r#return::Return as FunctionReturn;
 pub use self::context::pointer::Pointer;
 pub use self::context::r#loop::Loop;
 pub use self::context::traits::address_space::IAddressSpace;
 pub use self::context::traits::evmla_data::IEVMLAData;
+pub use self::context::traits::evmla_function::IEVMLAFunction;
 pub use self::context::value::Value;
 pub use self::context::IContext;
 pub use self::debug_config::ir_type::IRType as DebugConfigIR;
@@ -26,9 +30,6 @@ pub use self::eravm::build_assembly_text as eravm_build_assembly_text;
 pub use self::eravm::context::address_space::AddressSpace as EraVMAddressSpace;
 pub use self::eravm::context::build::Build as EraVMBuild;
 pub use self::eravm::context::evmla_data::EVMLAData as EraVMContextEVMLAData;
-pub use self::eravm::context::function::block::evmla_data::EVMLAData as EraVMFunctionBlockEVMLAData;
-pub use self::eravm::context::function::block::Block as EraVMFunctionBlock;
-pub use self::eravm::context::function::evmla_data::EVMLAData as EraVMFunctionEVMLAData;
 pub use self::eravm::context::function::intrinsics::Intrinsics as EraVMIntrinsicFunction;
 pub use self::eravm::context::function::llvm_runtime::LLVMRuntime as EraVMLLVMRuntime;
 pub use self::eravm::context::function::runtime::deploy_code::DeployCode as EraVMDeployCodeFunction;
@@ -74,9 +75,6 @@ pub use self::eravm::WriteLLVM as EraVMWriteLLVM;
 pub use self::evm::context::address_space::AddressSpace as EVMAddressSpace;
 pub use self::evm::context::build::Build as EVMBuild;
 pub use self::evm::context::evmla_data::EVMLAData as EVMContextEVMLAData;
-pub use self::evm::context::function::block::evmla_data::EVMLAData as EVMFunctionBlockEVMLAData;
-pub use self::evm::context::function::block::Block as EVMFunctionBlock;
-pub use self::evm::context::function::evmla_data::EVMLAData as EVMFunctionEVMLAData;
 pub use self::evm::context::function::intrinsics::Intrinsics as EVMIntrinsicFunction;
 pub use self::evm::context::function::runtime::entry::Entry as EVMEntryFunction;
 pub use self::evm::context::function::vyper_data::VyperData as EVMFunctionVyperData;
